Adding functions to update the database
Added functions update_departments and update_rooms. These functions
will create new Rooms or Departments in the database if they are not
there already.
Modified update_courses, making the Course object creation include the
new fields.

Things to fix: must create the correct Department object to be used in
the construction of the Course object. This can only be done correctly
when we know what should be used to query the database instead of the
id.