More uploads
Added upload for programs:
- Department programs must receive csv files containing the program name
and code
-Student programs refer to the programs of students enroled in a
specific course (name of the file is the course code). The file contains
utorids and program codes, as described in the new requirements.
-Program requirements: csv files named after the program code. They
contain tuples of formar [course-code, type-of-requirement], where
type-of-req is Mandatory, Writing or Stream (these were taken from the
CMS website)
-Added csv files to test these.
-Made the necessary additions to upload.html and script.js to include
these.